fix(ios): 补齐 K/N 2.3 的 BetaInteropApi opt-in，恢复图片解码失败显式抛异常
- iosMain 8 个文件补 @OptIn(BetaInteropApi::class)：K/N 2.3.20 的 stdlib 把该标记打在 @ObjCAction、
  ObjCObjectVar / .value、ObjCClassOf<T>.create(...) 这几类声明上，用到就必须显式 opt-in（伴生类方法本身
  不受约束，故 UIApplication.sharedApplication 等处不加）
- TextCharsetCodec.ios.kt：NSString.create 的 as String 保留并加 @Suppress("USELESS_CAST")，注释写明这是
  假告警 —— K/N 的 NSString↔String 映射只骗过了告警检查，删掉转换就无法当 String? 返回
- IosImageOps.decode：改用 UIImage.imageWithData（K/N 里返回 UIImage?）恢复"解码失败抛
  IllegalArgumentException"的契约。原先的 UIImage(data:) 是 @ObjCConstructor，静态类型非空，对它判空是
  死代码，nil 只能等 K/N 运行时抛空指针；上一版直接删掉守卫则连异常都没了
- IosImagePreview.toUIImage：同样改用 imageWithData，让"非图片字节返回 null"的注释与实际一致，不再靠调用方
  runCatching 把运行时异常当控制流
- ChineseUtils.ios.kt：删 CFStringTransform 返回值的多余 as String?
- IosSystemTtsEngine：AVSpeechUtteranceMinimum/MaximumSpeechRate 本就是 Float，去掉 .toFloat()
- LocalDate.native.kt：toYearMonthDay 直接读属性，删 LocalDateImpl 向下转型
- TextChapterSharedExt.getNeedReadAloud：StringBuilder.substring 已返回 String，删尾部 .toString()
- AlphanumComparator：形参 s1/s2 → a/b，避免与局部 s1Length/s2Length 混读

+++ shared/src/iosMain/kotlin/io/legado/app/help/tts/IosSystemTtsEngine.kt
@@ -216,8 +216,8 @@ class IosSystemTtsEngine : SystemTtsEngine {
             // 语速映射: speechRate ∈ [0.5, 2.0] → avRate ∈ [0.25, 1.0]
             // 公式: avRate = speechRate * 0.5 (1.0x → 0.5 defaultRate, 2.0x → 1.0 maxRate)
             rate = (rateMultiplier * 0.5f).coerceIn(
-                AVSpeechUtteranceMinimumSpeechRate.toFloat(),
-                AVSpeechUtteranceMaximumSpeechRate.toFloat()
+                AVSpeechUtteranceMinimumSpeechRate,
+                AVSpeechUtteranceMaximumSpeechRate
             )
         }
         synthesizer.speakUtterance(utterance)
